Nightly releaser (#38)
## Nightly Releaser

Workflow to uplift artifacts from nightly builds for `tt-forge-fe` and
`tt-torch`. `tt-xla` will be added later.

See nightly tags here.
https://github.com/tenstorrent/tt-forge/tags

### How it works

The workflow is a cron job that runs at 06:00 UTC daily and offers a
work dispatch to testing and re-runs.

It starts by calling the `uplift-artifacts` action, which will go
through the last 20 runs of a workflow's runs that uses a filter (e.g.,
repo, branch, workflow, completion status, etc.). After a workflow run
candidate is determined, it will download artifacts that match your glob
filter.

Then it calls ' Nightly release pruner` and keeps a max of 30 nightly
releases. Once that threshold is reached, the oldest nightly release
will be removed. This is used to keep our release page clean.

### Testing
Testing was done with the `on: push:` since the workflow needs to be
committed to the main for `work_dispatch` inputs to work.

#### Find a workflow in the main branch and get artifacts.

https://github.com/tenstorrent/tt-forge/actions/runs/13557769223/job/37895276983#step:4:82

#### Failed if no workflows on main completed successfully.

https://github.com/tenstorrent/tt-forge/actions/runs/13557559820/job/37894681838#step:3:101

#### Nightly release pruning (with limit 2)

https://github.com/tenstorrent/tt-forge/actions/runs/13555505330/job/37888795115#step:7:8