Workflow conclusion can be reflected through job conclusion (#85)
New refurbished nightly run for tt-forge-fe runs all available tests
including those known to fail.
However, the release is not dependent on these tests, even if failed we
claim that build is successful.
One of the most requested features on GitHub is "allow-failure" but
still it is not available.
So, there is fail-notify job that summarize nightly run and decide
whether it is successful or failed (and send notification on slack if
failed). Result of this job should be used as success mark for release
instead of whole workflow result.