Download artifact action (#111)
Default GitHub download-artifact action sometime fails in very odd way -
it doesn't download full artifact but action is successful. As we use
tar so we can preserve file attributes, untar step fails as archive is
damaged.
This happened occasionally, but lately it happens daily and became real
problem.
So, we designed our own download-action that will download and untar
artifact automatically. But If something fails, this action will retry
download the specified number of times with a wait time in between.
It is already working in tt-mlir, but now we had requests from other
repos also.