Perfomance benchmark test filter (#261)
Added filtering by project and test name for performance benchmarks so
it only launches test that fulfill the given conditions. Job matrix is
moved to `perf-bench-matrix.json` file.
Filtering support is also added to performance benchmark external
trigger run. Triggered runs are named so it is easier to identify it.
Perf benchmark has option to download and install latest wheel from
main. Install-wheel action is upgraded.
Running of perf benchmarks when docker is built is commented out due to
low number of runners (one) and old versions of tt-forge-fe released.