Enable device perf with dumped flatbuffer in tt-xla and ffe (#299)
For all frontends, device perf is calculated by compiling dumped ttir
and running ttrt perf on flatbuffer.
To avoid compiling twice, and to have same mlir_config flags that are
set in the frontend, ttrt perf will run on a flatbuffer that FFE and
tt-xla dump when compiling.

Also, added ttir from tt-torch and FFE to perf report.