bug: stash readme doc outside git path (#351)
The bug is caused by the git tag annotated checkout, which wipes the git
path clean and removes the generated release doc. This issue did not
surface in workflow draft testing because we do not check out other
repos during that test. The fix is to store the doc in the temp path of
the runner

```bash
⚠️ Unexpected error fetching GitHub release for tag refs/heads/main: Error: ENOENT: no such file or directory, open '/home/runner/work/tt-forge/tt-forge/release/docs/readme'
Error: ENOENT: no such file or directory, open '/home/runner/work/tt-forge/tt-forge/release/docs/readme'
```

https://github.com/tenstorrent/tt-forge/actions/runs/16929016574/job/47971430850#step:9:117

We can't change the Git path checkout due to the issues documented in
the `For the git checkout issue` section.
https://github.com/tenstorrent/tt-forge/issues/192