Performance tests custom ttrt wheel download (#370)
#### Issue
In the performance tests we were always downloading the ttrt wheel from
the HEAD of the main branch.
That caused issues because the frontend wheel could be from a different
commit and then compatibility is problematic.

#### Solution
With the new approach, we always download the ttrt wheel from the same
commit as the mlir version used by the frontend wheel.