Use CIv2 CPU runners for building the mega docker image & minor fixes (#389)
#### Issue 1
The mega docker image build step has started throwing an out of disk
space error (i.e. `ERROR: Could not install packages due to an OSError:
[Errno 28] No space left on device`)

The job has been changed to use the CIv2 CPU runners because of more
disk space in total.

#### Issue 2
When frontends use tt-mlir version with failed on-push workflow the ttrt
won't be downloaded.

This PR changes this and allows the ttrt wheel to be downloaded from the
on-push workflow, irrespective of the workflow's status.

#### Issue 3
In the `Install ttrt` step the `apt-get install` fails with `Unable to
fetch some archives`

Do a `apt-get update` before `apt-get install` to resolve the missing
archives fetch fail.

#### Improvement
Make the behaviour more clear in the syntax (if syntax and regex
syntax).

### Tests
The nightly release and tests workflow triggered on this PR:
[workflow](https://github.com/tenstorrent/tt-forge/actions/runs/17202995236).