Fix the download artifact action (#416)
The download artifact action didn't download a specific artifact.

The problem was that we fetch the list of artifacts from a specific
workflow, but the default page size for that endpoint is 30.
So if a workflow has more than 30 artifacts, the api call could
potentially not include the specific artifact we want in the response
and the action would not work like expected.

Fetch 100 artifacts from the specific workflow instead of the default
30.