Fix tt-xla perf benchmark tests dependency issue (#431)
#### Problem
Some tt-xla performance benchmark tests were installing `timm` in the
`Install system and python dependencies` step and that caused the
`torch` version to be bumped to 2.8.0.
That caused tests to fail because tt-xla wheel requires `torch==2.7.0`.

#### Solution
Specify `torch==2.7.0` in the `Install system and python dependencies`
step to ensure that the version won't be changed.