update relative link to point to the right folder (#460)
A page was moved from another repo that used a relative link to point at
information. The link did not work in the new location.

This pull request is to update the link in the README.md file to point
to the benchmarking section of the getting_started.md page for TT-Forge.