Adjustments to bge-m3 encode demo (#478)
### Ticket
#1196 

### Problem description
The `encode` function provided by BAAI/bge-m3 contains significant
overhead that is unnecessary and slows down the model.

### What's changed
Implemented a custom version of the `encode` function that tokenizes the
input sentences, runs the `model` on device and then performs the
expected output processing. The demo was also moved from the tt-torch
folder to the tt-xla folder as it uses the xla_backend.

### Checklist
- [x] New/Existing tests provide coverage for changes