Fix performance tests regression check for frontend uplifts (#501)
### Problem
When uplifting tt-mlir in tt-forge-fe we are checking if the changes
have introduces regressions in our performance tests.
We used to do that by comparing the results from the Nigthly workflow to
the new ones.
Due to the changes in tt-forge's Performance Benchmark workflow this has
stopped working properly.

### Solution
Change the parameters for filtering out the desired workflow whose
results we use to compare to the new ones.

### Test
Here are Performance Benchmark tests triggered with a regression check
from tt-forge-fe:
[workflow](https://github.com/tenstorrent/tt-forge-fe/actions/runs/18305865047)