Add arch parameter when checking for perf regressions (#622)
### Problem
The check for performance regressions compares the "old" perf result vs
the "current" perf result.
When fetching the "old" perf result we do not specify whether it's from
a test ran on a p150 or a n150 card.

### Solution
An additional parameter `arch` is sent when querying the db for perf
data.
It has two values `blackhole` if a test was ran on a p150 card or
`wormhole` if a test was ran on a n150 card.