Update tt-xla resnet demo postprocessing (#731)
The resnet demo is currently failing in demo tests since a change was
made to the post processing function in tt-forge-models. Updating the
post processing call to be `loader.output_postprocess(co_out=output)`
fixes this issue.

Implementing the change suggested in
https://github.com/tenstorrent/tt-forge/issues/685 (don't depend on
tt-forge-models for demos) would help to avoid similar issues in the
future.