Release pipeline fixes (#744)
There was a regression in the release pipeline ( filed @
https://github.com/tenstorrent/tt-xla/issues/2609 )

This PR was modified to address 2/3 regressions

For Albert, to fix the P0 ASAP, I'm modifying the postprocessing to be
done entirely on host. Fix of the underlying bug is coming later.
For OPT demo, this was a known failure caught on our nightly, a
consteval change causes out of memory issues, filed under
https://github.com/tenstorrent/tt-mlir/issues/5675 .